intermodal start/dest: try fix type overflow

+++ src/routing/journey.cc
@@ -24,7 +24,7 @@ void journey::leg::print(std::ostream& out,
                              },
                              [&](offset const x) {
                                indent(out, n_indent);
-                               out << "MUMO (id=" << static_cast<int>(x.type_)
+                               out << "MUMO (id=" << x.type_
                                    << ", duration=" << x.duration().count()
                                    << ")\n";
                              }},
